Substantially refactor almost everything.
This commit changes the primary top-level interface to be much
simpler, not requiring clients to know if they have `splice(2)`
available or not (a critical failure of the last API making it not
very nice to use.)

It refactors the Linux code all back into its own module, and the
portable interface behind it's own module. The top-level module
is instead just a very thin wrapper around the selected API
to make it convenient.

I also did some renaming and other general cleanup.

Signed-off-by: Austin Seipp <mad.one@gmail.com>