Rewrite splice(2) loop to not use blocking sockets.
GHC's event manager depends on I/O occuring in a non-blocking manner
because of the way thread scheduling works. This means the prior event
loop would clog up all running lightweight threads on some capability
when we needed to splice data.

Instead we just loop for EAGAIN when we call splice.

Also, close sockets in the splice(2) event loop when finished.

Signed-off-by: Austin Seipp <mad.one@gmail.com>